Introduce build system for roll20 Automatically builds up a roll20 compatible script by following the require chain. Seems to work ok
