+++ 5eShapedScripts/gulpfile.js
@@ -1,7 +1,7 @@
 const gulp = require('gulp');
 const mocha = require('gulp-mocha');
 const jshint = require('gulp-jshint');
-
+var webpack = require('webpack-stream');
 gulp.task('default', function() {
   // place code for your default task here
 });
@@ -12,6 +12,13 @@ gulp.task('lint', function () {
     return gulp.src('./components/*.js')
         .pipe(jshint())
         .pipe(jshint.reporter('default'));
+});
+
+gulp.task('make-roll20-js', function () {
+    'use strict';
+    return gulp.src('lib/entry-point.js')
+        .pipe(webpack(require('./webpack.config.js')))
+        .pipe(gulp.dest('./'));
 });
 
 gulp.task('test', function() {
