+++ 5eShapedScripts/lib/parser.js
