+++ 5eShapedScripts/package.json
@@ -2,7 +2,7 @@
   "name": "5e_shaped_scripts",
   "version": "0.1.0",
   "description": "Scripts to support the 5e Shaped Character Sheet on Roll20.net",
-  "main": "5eShapedScript.js",
+  "main": "lib/shaped-script.js",
   "dependencies": {
     "underscore": "^1.8.3"
   },
@@ -10,13 +10,15 @@
     "bluebird": "^3.3.4",
     "chai": "^3.5.0",
     "chai-as-promised": "^5.2.0",
-    "glob": "latest",
+    "glob": "^7.0.3",
     "gulp": "^3.9.1",
     "gulp-jshint": "^2.0.0",
     "gulp-mocha": "^2.2.0",
     "jshint": "^2.9.1",
+    "json-loader": "^0.5.4",
     "mocha": "^2.4.5",
-    "should": "^8.2.2"
+    "should": "^8.2.2",
+    "webpack-stream": "^3.1.0"
   },
   "scripts": {
     "test": "gulp test"
