+++ 5eShapedScripts/resources/mmFormatSpec.json
