+++ 5eShapedScripts/test/data/ancientGoldDragon.json
