+++ 5eShapedScripts/test/data/ancientGoldDragon.txt
