+++ 5eShapedScripts/test/data/beholder.txt
