+++ 5eShapedScripts/test/data/deathKnight.txt
