+++ 5eShapedScripts/test/data/lich.json
