+++ 5eShapedScripts/test/data/lich.txt
