+++ 5eShapedScripts/test/data/priest.txt
