+++ 5eShapedScripts/test/data/swarmOfQuippers.txt
