+++ 5eShapedScripts/test/suite.js
@@ -7,9 +7,19 @@ var Promise = require('bluebird'); // jshint ignore:line
 var fs = require('fs');
 Promise.promisifyAll(fs);
 var glob = require('glob');
-var getParser = require('../components/parser');
-var logger = require('../components/logger').getLogger({config: {logLevel: 'WARN'}});
-var sanitise = require('../components/sanitiser');
+var getParser = require('../lib/parser');
+var logger = require('../lib/logger')({logLevel: 'WARN'});
+var sanitise = require('../lib/sanitise');
+
+/**
+ * @name readFileAsync
+ * @memberOf fs
+ */
+
+/**
+ * @name sync
+ * @memberOf glob
+ */
 
 
 describe('Monster Manual tests', function () {
@@ -22,17 +32,16 @@ describe('Monster Manual tests', function () {
 
     var parser;
 
-    before(function (done) {
-        fs.readFile('./components/mmFormatSpec.json', 'utf-8', function (err, specText) {
+    before(function () {
+        return fs.readFileAsync('./resources/mmFormatSpec.json', 'utf-8')
+            .then(function (specText) {
             var parsed = JSON.parse(specText);
             parser = getParser(parsed, logger);
-            done();
         });
     });
 
 
-    //noinspection JSUnresolvedFunction
-    var files = glob.sync('./tests/data/*.txt');
+    var files = glob.sync('./test/data/*.txt');
     files.forEach(function (file) {
         it('correctly parses ' + file.replace(/\.txt$/, ''), function () {
             //noinspection JSUnresolvedVariable
@@ -45,7 +54,6 @@ describe('Monster Manual tests', function () {
 
 function runTestForFile(parser, file) {
     'use strict';
-    //noinspection JSUnresolvedFunction
     return fs.readFileAsync(file, 'utf-8').then(function (statblockText) {
         return runParse(parser, statblockText);
     });
@@ -55,7 +63,6 @@ function getExpectedOutputForFile(file) {
     'use strict';
 
     var filename = file.replace(/\.txt$/, '.json');
-    //noinspection JSUnresolvedFunction
     return fs.readFileAsync(filename, 'utf-8')
         .then(function (data) {
             return JSON.parse(data);
@@ -68,7 +75,6 @@ function getExpectedOutputForFile(file) {
 function runParse(parser, statBlockText) {
     'use strict';
     try {
-        console.log(parser);
         return parser.parse(sanitise(logger, statBlockText));
     }
     catch (e) {
