+++ 5eShapedScripts/webpack.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+    module: {
+        loaders: [
+            {test: /\.json$/, loader: 'json'}
+        ]
+    },
+    output: {
+        path: __dirname,
+        filename: 'ShapedScripts.js'
+    },
+    externals: {
+        'underscore': '_'
+    }
+};
